backend/requirements.txt
fastapi
uvicorn
hyperon
google-generativeai
python-dotenv
pydantic
pyjwt
passlib
bcrypt